requirements.txt
apache-airflow==2.4.0
apache-airflow-providers-celery
apache-airflow-providers-common-sql
apache-airflow-providers-ftp
apache-airflow-providers-google
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-postgres
apache-airflow-providers-slack
apache-airflow-providers-sqlite
numpy
pandas
sklearn
lshashpy3
--extra-index-url https://download.pytorch.org/whl/cu116
torch
torchvision
torchaudio 
psycopg2
transformers
bentoml
pydantic
tqdm
kubernetes
docker

